just 2 libs: domain (formely app) and shell (formely ports + adapters) there is a pool of connection inited at launch and each api endpoint can pool a conn instead of initing one for each call
big fat commit, sorry about that
