+++ rust/.env
@@ -1,2 +1,3 @@
 DATABASE_URL=postgres://postgres:postgres@localhost:5432/postgres
+ROCKET_DATABASES='{postgres={url="postgres://postgres:postgres@localhost:5432/postgres",idle_timeout=120}}'
 SECRET="shh its a secret"
