--- rust/adapters/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "adapters"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-application = { path = "../application" }
-diesel = { version = "2.2.0", features = ["postgres"] }
-dotenvy = "0.15.7"
-postgis_diesel = { version = "2.4.1", features = ["serde"] }
-hmac = "0.12.1"
-sha2 = "0.10.8"
