+++ rust/domain/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "application"
+name = "domain"
 version = "0.1.0"
 edition = "2021"
