+++ rust/domain/src/lib.rs
