+++ rust/domain/src/usecases/delete_post.rs
