--- rust/ports/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "ports"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-application = { path = "../application" }
-adapters = { path = "../adapters" }
-rocket= { version = "0.5.1", features = ["json"] }
-serde = { version = "1.0", features = ["derive"] }
-jwt = "0.16.0"
