+++ rust/shell/.gitignore
