+++ rust/shell/diesel.toml
@@ -2,7 +2,7 @@
 # see https://diesel.rs/guides/configuring-diesel-cli
 
 [print_schema]
-file = "src/schema.rs"
+file = "src/db/schema.rs"
 custom_type_derives = ["diesel::query_builder::QueryId", "Clone"]
 import_types = ["diesel::sql_types::*", "postgis_diesel::sql_types::*"]
 filter = { except_tables = ["spatial_ref_sys"] }
