+++ rust/shell/src/auth.rs
@@ -1,4 +1,5 @@
 use super::error::Error;
+
 use jwt::{SignWithKey, VerifyWithKey};
 use rocket::{
     http::Status,
@@ -6,8 +7,18 @@ use rocket::{
     Request,
 };
 use std::collections::BTreeMap;
+use hmac::{Hmac, Mac};
+use sha2::Sha256;
+use dotenvy::dotenv;
+use std::env;
 
-use adapters::load_hmac_key;
+pub fn load_hmac_key() -> Result<Hmac<Sha256>, Error> {
+    dotenv()?;
+    let secret = env::var("SECRET")?;
+    let key: Hmac<Sha256> = Hmac::new_from_slice(secret.as_bytes())?;
+
+    Ok(key)
+}
 
 pub fn _sign_token(sub: String) -> Result<String, Error> {
     let key = load_hmac_key()?;
