+++ rust/shell/src/db/schema.rs
